+++ Flow/src/utopia/flow/util/StringExtensions.scala
@@ -25,13 +25,13 @@ object StringExtensions
 		/**
 		 * @return Words that belong to this string. <b>This includes all non-whitespace characters but not newline characters</b>
 		 */
-		def words = s.linesIterator.toVector.flatMap { _.split(" ").toVector.map { _.trim }.filter { _.nonEmpty } }
+		def words = s.linesIterator.toVector.flatMap { _.split(" ").toVector
+			.map { _.trim }.filter { _.nonEmpty } }
 		
 		/**
 		 * @return The first word in this string (may include any characters except whitespace)
 		 */
 		def firstWord = untilFirst(" ")
-		
 		/**
 		 * @return The last word in this string (may include any characters except whitespace)
 		 */
@@ -55,6 +55,10 @@ object StringExtensions
 		  * @return A copy of this string surrounded with quotation marks (")
 		  */
 		def quoted = "\"" + s + "\""
+		/**
+		  * @return A copy of this string where the first character is in lower case
+		  */
+		def uncapitalize = if (s.isEmpty) s else s"${s.head.toLower}${s.drop(1)}"
 		
 		/**
 		  * @param range A range
@@ -109,7 +113,8 @@ object StringExtensions
 		 * @param more More strings
 		 * @return Whether this string contains all of the provided sub-strings (case-sensitive)
 		 */
-		def containsAll(first: String, second: String, more: String*): Boolean = containsAll(Vector(first, second) ++ more)
+		def containsAll(first: String, second: String, more: String*): Boolean =
+			containsAll(Vector(first, second) ++ more)
 		
 		/**
 		 * @param strings A number of strings